fix: comprehensive review findings (B blocker + C/E nits)
Reviewer B (Grade B+, blocker): widen SplitPane API from f32 to f64 to
satisfy API_DESIGN.md Rule 2 (floats are f64 everywhere). Touched fields:
- SplitPaneState::{ratio, min_ratio, new, with_min_ratio, set_ratio}
- SplitPaneResponse::ratio
- DEFAULT_SPLIT_MIN_RATIO
- KEY_STEP, RATIO_GROW_SCALE (split.rs internal constants)
- delta accumulator in consume_split_pane_keys
- Drag-math casts (rel as f32 → f64::from(rel)) in consume_split_pane_drag

Reviewer C (Grade A-, two warnings + nits):
- scripts/ghostty_demos.sh — add v020_test_utils to FEATURES_PRESET
  (was asymmetric with v020_perf_audit which was already included)
- examples/v020_dx_shortcuts.rs — replace .gap(1) / .p(1) / .mt(1)
  magic literals with theme.spacing.xs()
- examples/v020_regression_panel.rs — replace .p(2) magic literal with
  theme.spacing.sm()

Reviewer E (Grade A, two nits):
- CHANGELOG: remove redundant HighlightRange::single entry from Breaking
  (it's already mentioned inline in the #235 entry as an additive alias)
- CHANGELOG: rename "f32 → f64 unification" entry to cover the SplitPane
  widening too, with a note about the style/color.rs f32 internals being
  the documented Rule 2 graphics-math exception

All quality gates pass after fixes:
- cargo fmt -- --check OK
- cargo check --all-features OK
- cargo clippy --all-features -- -D warnings OK
- cargo test --all-features OK (1080+ tests)
- cargo check --examples --all-features OK

+++ src/context/widgets_display/split.rs
@@ -7,13 +7,13 @@
 use super::*;
 
 /// Keyboard step applied to `state.ratio` per arrow-key press when the handle is focused.
-const KEY_STEP: f32 = 0.05;
+const KEY_STEP: f64 = 0.05;
 
 /// Scale factor applied to the `[0.0, 1.0]` ratio to produce a `u16` flexbox
 /// `grow` weight. 1000 gives ~0.1% precision in pane sizes — finer than any
 /// terminal cell can render at typical widths, while staying well below
 /// `u16::MAX` so the two-pane sum can never overflow.
-const RATIO_GROW_SCALE: f32 = 1000.0;
+const RATIO_GROW_SCALE: f64 = 1000.0;
 
 /// Direction of the split. Internal helper — public API is the `split_pane`
 /// (horizontal) / `vsplit_pane` (vertical) entry points.
@@ -150,7 +150,7 @@ impl Context {
         orientation: SplitOrientation,
     ) {
         let mut consumed: Vec<usize> = Vec::new();
-        let mut delta = 0.0_f32;
+        let mut delta = 0.0_f64;
         for (i, key) in self.available_key_presses() {
             let is_left = matches!(key.code, KeyCode::Left);
             let is_right = matches!(key.code, KeyCode::Right);
@@ -248,7 +248,7 @@ impl Context {
                                         .x
                                         .saturating_sub(outer.x)
                                         .min(outer.width.saturating_sub(1));
-                                    rel as f32 / outer.width as f32
+                                    f64::from(rel) / f64::from(outer.width)
                                 }
                             }
                             SplitOrientation::Vertical => {
@@ -259,7 +259,7 @@ impl Context {
                                         .y
                                         .saturating_sub(outer.y)
                                         .min(outer.height.saturating_sub(1));
-                                    rel as f32 / outer.height as f32
+                                    f64::from(rel) / f64::from(outer.height)
                                 }
                             }
                         };
